+++ flow-maven-plugin/src/test/java/com/vaadin/flow/plugin/maven/BuildFrontendMojoTest.java
@@ -27,9 +27,11 @@ import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import org.apache.maven.model.Build;
 import org.apache.maven.plugin.AbstractMojo;
@@ -67,6 +69,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 public class BuildFrontendMojoTest {
+    public static final String TEST_PROJECT_RESOURCE_JS = "test_project_resource.js";
     @Rule
     public TemporaryFolder temporaryFolder = new TemporaryFolder();
 
@@ -74,6 +77,7 @@ public class BuildFrontendMojoTest {
     private File generatedFolder;
     private File nodeModulesPath;
     private File flowPackagPath;
+    private File projectFrontendResourcesDirectory;
     private String mainPackage;
     private String appPackage;
     private String webpackConfig;
@@ -102,6 +106,18 @@ public class BuildFrontendMojoTest {
         appPackage = new File(generatedFolder, PACKAGE_JSON).getAbsolutePath();
         webpackConfig = new File(npmFolder, WEBPACK_CONFIG).getAbsolutePath();
 
+
+        projectFrontendResourcesDirectory = new File(npmFolder,
+                "flow_resources");
+
+        Assert.assertTrue("Failed to create a test project resources",
+                projectFrontendResourcesDirectory.mkdirs());
+        Assert.assertTrue("Failed to create a test project file",
+                new File(projectFrontendResourcesDirectory,
+                        TEST_PROJECT_RESOURCE_JS).createNewFile());
+
+        ReflectionUtils.setVariableValueInObject(mojo, "frontendResourcesDirectory", projectFrontendResourcesDirectory);
+
         ReflectionUtils.setVariableValueInObject(mojo, "project", project);
         ReflectionUtils.setVariableValueInObject(mojo, "generatedFolder",
                 generatedFolder);
@@ -128,7 +144,7 @@ public class BuildFrontendMojoTest {
     }
 
     @After
-    public void teardown() {
+    public void teardown() throws IOException {
         if (FileUtils.fileExists(mainPackage)) {
             FileUtils.fileDelete(mainPackage);
         }
@@ -149,6 +165,30 @@ public class BuildFrontendMojoTest {
         when(project.getBuild()).thenReturn(buildMock);
         when(project.getRuntimeClasspathElements()).thenReturn(getClassPath());
         ReflectionUtils.setVariableValueInObject(mojo, "project", project);
+    }
+
+    @Test
+    public void should_copyProjectFrontendResources() {
+
+        List<File> initialFiles = gatherFiles(nodeModulesPath);
+        initialFiles.forEach(file -> Assert.assertFalse(String.format(
+                        "Test resource shouldn't exist before running mojo.",
+                        file), TEST_PROJECT_RESOURCE_JS.equals(file.getName())));
+        mojo.execute();
+
+        Set<String> projectFrontendResources = Stream
+                .of(projectFrontendResourcesDirectory.listFiles())
+                .map(File::getName).collect(Collectors.toSet());
+        List<File> filesInNodeModules = gatherFiles(nodeModulesPath);
+filesInNodeModules.removeAll(initialFiles);
+
+        Assert.assertEquals(
+                "All project resources should be copied into the node_modules",
+                projectFrontendResources.size(), filesInNodeModules.size());
+
+        filesInNodeModules.forEach(file -> Assert.assertTrue(String.format(
+                "Expected the copied file '%s' to be in the project resources",
+                file), projectFrontendResources.contains(file.getName())));
     }
 
     @Test
@@ -429,5 +469,21 @@ public class BuildFrontendMojoTest {
             }
         }
         return classPaths;
+    }
+
+    private List<File> gatherFiles(File root) {
+        if (root.isFile()) {
+            return Collections.singletonList(root);
+        } else {
+            File[] subdirectoryFiles = root.listFiles();
+            if (subdirectoryFiles != null) {
+                List<File> files = new ArrayList<>();
+                for (File subdirectoryFile : subdirectoryFiles) {
+                    files.addAll(gatherFiles(subdirectoryFile));
+                }
+                return files;
+            }
+            return Collections.emptyList();
+        }
     }
 }
